+++ src/scvi_de/__init__.py
@@ -1,42 +1,79 @@
 import warnings
+from pathlib import Path
 from typing import Literal
 
-import anndata as ad
+import muon as mu
+import numpy.typing as npt
 import pandas as pd
 import scvi
+from anndata import AnnData
+from loguru import logger
+from mudata import MuData
 from scipy.sparse import csr_matrix, issparse
 
+from scvi_de.logging import init_logger
 
+init_logger(verbose=3)
+
+
+# TODO: add logging
 # TODO: pass more options on
+# TODO: add an option to save the model IMMEDIATELY after creation
+# TODO: FUCKING HELL WRITE SOME UNIT TESTS!
 def scvi_de(
-    adata: ad.AnnData | None = None,
+    adata: AnnData | MuData | None = None,
     model: scvi.model.SCVI | scvi.model.TOTALVI | None = None,
+    save_model_path: Path | None = None,
+    overwrite_previous_model: bool = False,
+    protein_obsm_key: str = "prot",
+    protein_names_uns_key: str = "prot_names",
     groupby: str = "leiden",
-    layer: str | None = None,
+    layers: str | dict[str, str] | None = None,
+    protein_layer: str | None = None,
     modality: Literal["rna", "prot"] = "rna",
+    mudata_protein_modality: str = "prot",
+    mudata_rna_modality: str = "rna",
     compare_group: str | None = None,
-    reference_group: str = "rest",
+    reference_group: str | None = None,
     batch_correct: bool = False,
     batch_key: str | None = None,
-    dispersion: Literal["gene", "gene-batch", "gene-label", "gene-cell", "protein", "protein-batch", "protein-label"] | None = None,
+    dispersion: (
+        Literal[
+            "gene",
+            "gene-batch",
+            "gene-label",
+            "gene-cell",
+            "protein",
+            "protein-batch",
+            "protein-label",
+        ]
+        | None
+    ) = None,
     gene_likelihood: Literal["nb", "zinb", "poisson"] = "nb",
     remove_outliers: bool = False,
-    inplace: bool = False,
     return_df: bool = True,
     return_model: bool = True,
     lfc_use: Literal["lfc_mean", "lfc_median", "lfc_max", "lfc_min"] = "lfc_mean",
+    num_devices: str | int = "auto",
     **kwargs,
-) -> dict[str, pd.DataFrame | ad.AnnData | scvi.model.SCVI | scvi.model.TOTALVI] | None:
+) -> dict[str, pd.DataFrame | AnnData | scvi.model.SCVI | scvi.model.TOTALVI] | None:
     """Perform differential expression analysis using the scVI suite of tools
 
     Parameters
     -------
-    adata : anndata.AnnData
-        Object to model and perform differential expression on. Either an AnnData or model object need to be passed;
+    adata : :class:`~anndata.AnnData` | :class:`~MuData`
+        Object to model and perform differential expression on. Either an AnnData/MuData or model object need to be passed;
         if both are passed, AnnData will be ignored.
     model : :class:`~scvi.model.SCVI`, :class:`~scvi.model.TOTALVI`
         A previously trained model.  If both an AnnData and model object are passed, the AnnData object will be ignored
         and the model will be used.
+    save_model_path : :class:`~pathlib.Path`
+        Path to a folder in which to save the model and associated anndata object. This value will be used to create a
+        folder within which will be placed 'model.pt' and 'adata.h5ad' files. This WILL overwrite any e
+    protein_obsm_key: str | None, default=None
+
+    protein_names_uns_key: str | None, default=None
+
     groupby : str, default="leiden"
         Column in `obs` to use when performing comparisons.
     layer : str
@@ -75,8 +112,6 @@ def scvi_de(
 
     remove_outliers : bool, default=False
         Whether to filter outlier cells with `:meth:~scvi.model.base.DifferentialComputation.filter_outlier_cells()`
-    inplace : bool, default=False
-        Should any changes be performed directly on the provided AnnData object. If not, a copy is made.
     return_df : bool, default=True
         Should a dataframe of results be returned?
     return_model : bool, default=True
@@ -86,7 +121,8 @@ def scvi_de(
 
     Return
     ------
-        dict of {str: pd.DataFrame | ad.AnnData | scvi.model.SCVI | scvi.model.TOTALVI}
+        dict of {str: pd.DataFrame | scvi.model.SCVI | scvi.model.TOTALVI}
+        Original object is modified in place, with results placed in adata.uns["rank_genes_groups"]
     """
     if dispersion is None:
         match modality:
@@ -97,58 +133,178 @@ def scvi_de(
             case _:
                 msg = "Dispersion was not indicated and cannot be figured out based on the type of data to be analyzed"
                 raise ValueError(msg)
-    if dispersion not in ("gene", "gene-batch", "gene-label", "gene-cell", "protein", "protein-batch", "protein-label"):
-        msg = f"{dispersion} is not a valid option for dispersion"
-        raise ValueError(msg)
-        
-    if adata and model:
-        msg = (
-            "Passing and Anndata object and scVI model is confusing. "
-            "If you wish to regenerate the model, pass just an Anndata object. "
-            "For now, using the previous model in the interests of time..."
+
+    # several check to make sure we don't waste a lot of time on something that will later fail
+    logger.debug("performing checks")
+    perform_checks(
+        adata,
+        model,
+        save_model_path,
+        overwrite_previous_model,
+        protein_obsm_key,
+        modality,
+        mudata_protein_modality,
+        mudata_rna_modality,
+        dispersion,
+    )
+
+    # scvi sometimes takes MuData objects, sometimes not.  I don't think TotalVI does, so we need to reform the data
+    # both scVI and TotalVI models assume RNA data in X or the indicated layer so need this in either case
+    tmp = (
+        process_mudata(
+            adata,
+            layers,
         )
-        warnings.warn(msg, UserWarning, stacklevel=2)
-    elif adata:
+        if isinstance(adata, MuData)
+        else process_anndata(adata, layers)
+    )
+
+    if adata and not model:
         model = create_model(
-            adata=adata,
-            layer=layer,
+            adata=tmp,
+            layers=layers,
             batch_key=batch_key,
-            inplace=inplace,
             dispersion=dispersion,
             gene_likelihood=gene_likelihood,
             modality=modality,
+            protein_obsm_key=protein_obsm_key,
+            protein_names_uns_key=protein_names_uns_key,
+            num_devices=num_devices,
             **kwargs,
         )
-    else:
+    elif (adata is None) and (model is None):
         msg = "Either an AnnData or scVI model needed and both are missing."
         raise ValueError(msg)
 
-    adata_copy = model.adata
-    latent = model.get_latent_representation()
-    adata_copy.obsm["X_scVI"] = latent
-    if compare_group:
-        de_change = one_vs_others(
-            adata=adata_copy,
-            model=model,
-            groupby=groupby,
-            current_group=compare_group,
-            reference_group=reference_group,
-            batch_correct=batch_correct,
-            remove_outliers=remove_outliers,
-        )
-    else:
-        de_change = {
-            i: one_vs_others(
-                adata=adata_copy,
-                model=model,
-                groupby=groupby,
-                current_group=i,
-                batch_correct=batch_correct,
-                remove_outliers=remove_outliers,
-            )
-            for i in adata_copy.obs[groupby].unique()
-        }
+    if save_model_path:
+        model.save(save_model_path, overwrite=overwrite_previous_model, save_anndata=True)
 
+    de_change = diff_expr_test(
+        adata=model.adata,
+        model=model,
+        groupby=groupby,
+        current_group=compare_group,
+        reference_group=reference_group,
+        batch_correct=batch_correct,
+        remove_outliers=remove_outliers,
+    )
+
+    model.adata.obsm["X_scVI"] = model.get_latent_representation()
+
+    return_dict = {}
+
+    if isinstance(model.adata, AnnData):
+        model.adata.uns["rank_genes_groups"] = process_deg_results(de_change, lfc_use)
+    elif isinstance(model.adata, MuData):
+        for _ in model.adata.mod:
+            if isinstance(de_change, dict):
+                model.adata[_].uns["rank_genes_groups"] = process_deg_results(
+                    {i: de_change[i].loc[model.adata[_].var_names, :] for i in de_change}, lfc_use
+                )
+
+    if return_df:
+        return_dict["deg_df"] = de_change
+    if return_model:
+        return_dict["model"] = model
+
+    return return_dict
+
+
+def perform_checks(
+    adata,
+    model,
+    save_model_path,
+    overwrite_previous_model,
+    protein_obsm_key,
+    modality,
+    mudata_protein_modality,
+    mudata_rna_modality,
+    dispersion,
+):
+    # Do we have a mismatch of modality and dispersion options?
+    if ((modality == "rna") and ("protein" in dispersion)) or ((modality == "prot") and ("gene" in dispersion)):
+        msg = (
+            f"The argument you passed indicate that this is {modality} data, but then you also indicated a "
+            f"{dispersion}. These are incompatible. Please try again."
+        )
+        raise ValueError(msg)
+
+    if save_model_path and (save_model_path.exists() and not overwrite_previous_model):
+        msg = (
+            f"It appears that a folder exists at {save_model_path}, but `overwrite_previous_model` is `False`. "
+            f"Make sure that is correct before long and expensive calculations are carried out!"
+        )
+        warnings.warn(msg, UserWarning, stacklevel=2)
+
+    if isinstance(adata, MuData):
+        for _ in (mudata_protein_modality, mudata_rna_modality):
+            if _ not in adata.mod:
+                msg = (
+                    f"{_} does not appear to be the name of a modality in the passed object. "
+                    f"The available modalities are {','.join(adata.mod)}."
+                )
+                raise ValueError(msg)
+    if (protein_obsm_key is None) and ("prot" in modality.lower()) and isinstance(adata, AnnData):
+        msg = (
+            "A name for the key where protein data in adata.obsm was not provided, but the modality selected is protein. "
+            "If assessing protein data, it is recommended to to either use a MuData object or place transcript counts "
+            "in adata.X or a layer in adata and the protein counts in adata.obsm[protein_obsm_key]. Results WILL differ."
+        )
+        warnings.warn(msg, UserWarning, stacklevel=2)
+    elif isinstance(adata, AnnData) and ("prot" in modality.lower()):
+        msg = (
+            "If assessing protein data, it is recommended to to either use a MuData object or place transcript counts "
+            "in adata.X or a layer in adata and the protein counts in adata.obsm[protein_obsm_key]. Results WILL differ."
+        )
+        warnings.warn(msg, UserWarning, stacklevel=2)
+
+    if dispersion not in (
+        "gene",
+        "gene-batch",
+        "gene-label",
+        "gene-cell",
+        "protein",
+        "protein-batch",
+        "protein-label",
+    ):
+        msg = f"{dispersion} is not a valid option for dispersion"
+        raise ValueError(msg)
+
+    if adata and model:
+        msg = (
+            "When both an Anndata/MuData object and scVI model are passed to scvi_de, the existing model is used for "
+            "comparisons and is not regenerated. Instead, the differential expression results are written to "
+            "the provided Anndata/MuData object in adata.uns['rank_genes_groups']."
+        )
+        warnings.warn(msg, UserWarning, stacklevel=2)
+    if model:
+        msg = (
+            "Reusing the provided model. If you want to add the results from the current comparisons to your "
+            "AnnData/Mudata object, run `adata.uns['rank_genes_groups'] = process_deg_results(degs['deg_df'])` "
+            "or the MuData equivalent."
+        )
+        warnings.warn(msg, UserWarning, stacklevel=2)
+
+
+def process_deg_results(
+    de_change, lfc_use: Literal["lfc_mean", "lfc_median", "lfc_std", "lfc_min", "lfc_max"] = "lfc_mean"
+):
+    """
+    Reformat the dataframe(s) returned by `scvi.model.MODEL.differential_expression()` to a form matching
+    that returned by `scanpy.tl.rank_genes_groups()`.
+
+    Parameters
+    ----------
+    de_change : :class:`pd.DataFrame` | dict[str, :class:`pd.DataFrame`]
+        A dataframe for a single comparison or dictionary of dataframes for multiple comparisons, as returned by
+        `scvi.model.MODEL.differential_expression()`
+    lfc_use : :class:`Literal["lfc_mean", "lfc_median", "lfc_std", "lfc_min", "lfc_max"]`, default="lfc_mean"
+        Which log-fold change column in the dataframes to use for fold change values
+
+    Return
+    ------
+    Same as that returned by `scanpy.tl.rank_genes_groups()`
+    """
     names = {
         i: (de_change[i]["proba_de"] * de_change[i][lfc_use] * de_change[i]["non_zeros_proportion1"])
         .sort_values(ascending=False)
@@ -173,7 +329,7 @@ def scvi_de(
         index=False
     )
 
-    de_dict = {
+    return {
         "params": {
             "groupby": "leiden_wnn_0.9",
             "reference": "rest",
@@ -190,28 +346,51 @@ def scvi_de(
         "logfoldchanges": lfc_rec,
     }
 
-    return_dict = {}
-    if inplace:
-        adata.uns["rank_genes_groups"] = de_dict
-    else:
-        adata_copy.uns["rank_genes_groups"] = de_dict
-        return_dict["adata"] = adata_copy
 
-    if return_df:
-        return_dict["deg_df"] = de_change
-    if return_model:
-        return_dict["model"] = model
-    return return_dict
+def process_mudata(
+    mudata: MuData,
+    layers: dict[str, str] | None = None,
+) -> MuData:
+    for mod in mudata.mod:
+        if layers and (mod in layers and not is_integer_array(mudata[mod].layers[layers[mod]])):
+            msg = (
+                f"TOTALVI require raw, non-normalized counts and it looks like"
+                f"layers[{mod}] in the {mod} modality is not all integers."
+            )
+            raise ValueError(msg)
+        # not using a layer, so check X and if it isn't all integers, try pulling from raw if possible
+        elif not is_integer_array(mudata[mod].X):
+            if mudata[mod].raw and is_integer_array(mudata[mod].raw.X):
+                mudata[mod].X = mudata[mod].raw[mudata[mod].obs_names, mudata[mod].var_names].X.copy()
+                mudata[mod].X = mudata[mod].X.toarray() if issparse(mudata[mod].X) else mudata[mod].X
+                # no? fuck it. error
+            else:
+                msg = (
+                    f"TOTALVI requires raw, non-normalized counts and it looks a raw "
+                    f"integer array for the {mod} modality could not be found in X nor the raw slot."
+                )
+                raise ValueError(msg)
+
+    return mudata
 
 
 def create_model(
-    adata: ad.AnnData,
-    layer: str | None = None,
+    adata: AnnData | MuData,
+    layers: str | dict[str, str] | None = None,
     modality: Literal["rna", "prot"] = "rna",  # add more options and we validate input arguments
+    mudata_rna_modality: str = "rna",
+    mudata_protein_modality: str = "prot",
     batch_key: str | None = None,
-    inplace: bool = False,
+    protein_obsm_key: str | None = None,
+    protein_names_uns_key: str | None = None,
     dispersion: Literal[
-        "gene", "gene-batch", "gene-label", "gene-cell", "protein", "protein-batch", "protein-label"
+        "gene",
+        "gene-batch",
+        "gene-label",
+        "gene-cell",
+        "protein",
+        "protein-batch",
+        "protein-label",
     ] = "gene",
     gene_likelihood: Literal["nb", "zinb", "poisson"] = "nb",
     save_layer: str = "counts",
@@ -222,63 +401,79 @@ def create_model(
     early_stopping_monitor: Literal[
         "elbo_validation", "reconstruction_loss_validation", "kl_local_validation"
     ] = "elbo_validation",
+    num_devices: str | int = "auto",
     **kwargs,
 ) -> scvi.model.SCVI:
-    if ((modality == "rna") and ("protein" in dispersion)) or ((modality == "prot") and ("gene" in dispersion)):
-        msg = f"The argument you passed indicate that this is {modality} data, but then you also indicated a {dispersion}. These are incompatible. Please try again."
-        raise ValueError(msg)
+    # copying the object so that we don't fuckup the source.  Could maybe add back an option to not do this to save
+    # on memory
+    adata_copy = adata.copy()
 
-    adata_copy = adata if inplace else adata.copy()
-    if layer:
-        is_not_int = (
-            ((adata_copy.layers[layer].toarray() % 1) != 0).any()
-            if issparse(adata_copy.layers[layer])
-            else ((adata_copy.layers[layer] % 1) != 0).any()
-        )
-        if is_not_int:
-            msg = f"Both the scVI and TOTALVI models require raw, non-normalized counts and it looks like the {layer} layer in the passed object are normalized"
-            raise ValueError(msg)
-    else:
-        is_not_int = (
-            ((adata_copy.X.toarray() % 1) != 0).any() if issparse(adata_copy.X) else ((adata_copy.X % 1) != 0).any()
-        )
-        if is_not_int:
-            msg = "the scVI and TOTALVI models require raw, non-normalized counts and it looks like the passed object has been normalized"
-            raise ValueError(msg)
+    # identify the HVGs and subset on those
+    if isinstance(adata_copy, AnnData):
+        scvi.data.poisson_gene_selection(adata_copy, layer=layers)
+        adata_copy = adata_copy[:, adata_copy.var["highly_variable"]]
+    elif isinstance(adata_copy, MuData):
+        if "rna" in adata_copy.mod:
+            scvi.data.poisson_gene_selection(adata_copy["rna"], layer=layers)
+            mu.pp.filter_var(adata_copy["rna"], var="highly_variable")
 
-    scvi.data.poisson_gene_selection(adata_copy, layer=layer)
-
-    adata_copy = adata_copy[:, adata_copy.var["highly_variable"]]  # focus on selected genes
-
-    if layer:
+    if layers and isinstance(adata_copy, AnnData):
         adata_copy.layers["X"] = adata_copy.X.copy()
-        adata_copy.X = adata_copy.layers[layer].copy()
-    adata_copy.layers[save_layer] = csr_matrix(adata_copy.X.copy())  # converts to CSR format, preserve counts
+        adata_copy.X = adata_copy.layers[layers].copy()
+        adata_copy.layers[save_layer] = csr_matrix(adata_copy.X.copy())  # converts to CSR format, preserve counts
+    if isinstance(adata_copy, MuData):
+        for mod in adata_copy.mod:
+            adata_copy[mod].layers[save_layer] = csr_matrix(
+                adata_copy[mod].X.copy()
+            )  # converts to CSR format, preserve counts
 
+    # If we are working with RNA data, use scVI; for protein, use TotalVI
     match modality.lower():
         case "rna":
-            scvi.model.SCVI.setup_anndata(adata_copy, layer=layer, batch_key=batch_key)
+            scvi.model.SCVI.setup_anndata(adata_copy, layer=layers, batch_key=batch_key)
             model = scvi.model.SCVI(adata_copy, dispersion=dispersion, gene_likelihood=gene_likelihood)
         case "prot":
-            # assuming that we're working with the protein portion of a MuData object
-            if issparse(adata_copy.X):
-                adata_copy.obsm["prot"] = adata_copy.X.toarray().copy()
-            else:
-                adata_copy.obsm["prot"] = adata_copy.X.copy()
-            adata_copy.uns["prot_names"] = [f"{_}_x" for _ in adata_copy.var_names.to_list()]
-            scvi.model.TOTALVI.setup_anndata(
-                adata=adata_copy,
-                layer=save_layer,
-                batch_key=batch_key,
-                protein_expression_obsm_key="prot",
-                # there is currently (2024-04-16) a bug in TotalVI that causes a
-                # "cannot reindex on an axis with duplicate labels" error.
-                # See: https://github.com/scverse/scvi-tools/issues/2627
-                # I, however, think this is because gene expression should stay in adata.X
-                # and prot data should be stuffed in adata.obsm[prot_key]
-                protein_names_uns_key="prot_names",
+            # Two ways to set things up:
+            # if an AnnData object, we need data in adata.obsm[protein_obsm_key]
+            # and we need the feature names in adata.uns[protein_names_uns_key]
+            if isinstance(adata, AnnData):
+                if issparse(adata_copy.X):
+                    adata_copy.obsm[protein_obsm_key] = adata_copy.X.toarray().copy()
+                else:
+                    adata_copy.obsm[protein_obsm_key] = adata_copy.X.copy()
+                adata_copy.uns[protein_names_uns_key] = [f"{_}_x" for _ in adata_copy.var_names.to_list()]
+                scvi.model.TOTALVI.setup_anndata(
+                    adata=adata_copy,
+                    layer=save_layer,
+                    batch_key=batch_key,
+                    protein_expression_obsm_key=protein_obsm_key,
+                    # there is currently (2024-04-16) a bug in TotalVI that causes a
+                    # "cannot reindex on an axis with duplicate labels" error.
+                    # See: https://github.com/scverse/scvi-tools/issues/2627
+                    # I, however, think this is because gene expression should stay in adata.X
+                    # and prot data should be stuffed in adata.obsm[prot_key]
+                    protein_names_uns_key=protein_names_uns_key,
+                )
+            # things are much simpler if we have a MuData object
+            elif isinstance(adata, MuData):
+                if layers:
+                    protein_layer = layers[mudata_protein_modality] if mudata_protein_modality in layers else None
+                    rna_layer = layers[mudata_rna_modality] if mudata_rna_modality in layers else None
+                else:
+                    protein_layer = None
+                    rna_layer = None
+                scvi.model.TOTALVI.setup_mudata(
+                    mdata=adata_copy,
+                    batch_key=batch_key,
+                    rna_layer=rna_layer,
+                    protein_layer=protein_layer,
+                    modalities={"rna_layer": mudata_rna_modality, "protein_layer": mudata_protein_modality},
+                )
+            model = scvi.model.TOTALVI(
+                adata_copy,
+                protein_dispersion=dispersion,
+                gene_likelihood=gene_likelihood,
             )
-            model = scvi.model.TOTALVI(adata_copy, protein_dispersion=dispersion, gene_likelihood=gene_likelihood)
 
     model.train(
         check_val_every_n_epoch=check_val_every_n_epoch,
@@ -286,42 +481,68 @@ def create_model(
         early_stopping=early_stopping,
         early_stopping_patience=early_stopping_patience,
         early_stopping_monitor=early_stopping_monitor,
+        devices=num_devices,
         **kwargs,
     )
 
     return model
 
 
-def one_vs_others(
-    adata: ad.AnnData,
+def process_anndata(adata, layers) -> AnnData:
+    if layers:
+        for _ in layers:
+            if not is_integer_array(adata.layers[_]):
+                msg = (
+                    f"Both the scVI and TOTALVI models require raw, non-normalized counts and it looks like the {layers[_]} "
+                    f"layer in the passed object are normalized"
+                )
+                raise ValueError(msg)
+    elif not is_integer_array(adata.X):
+        msg = (
+            "the scVI and TOTALVI models require raw, non-normalized counts and it looks like the passed object has "
+            "been normalized"
+        )
+        raise ValueError(msg)
+    return adata
+
+
+def diff_expr_test(
+    adata: AnnData,
     model: scvi.model.SCVI,
     groupby: str,
-    current_group: str,
-    reference_group: str = "rest",
+    current_group: str | None = None,
+    reference_group: str | None = None,
     batch_correct: bool = False,
     remove_outliers: bool = False,
 ):
-    cell_idx1 = adata.obs[groupby] == current_group
-
-    # honestly, if we are doing one vs rest comparisons, we need to change this to 
-    if reference_group == "rest":
-        cell_idx2 = adata.obs[groupby] != current_group
-    else:
-        cell_idx2 = adata.obs[groupby] == reference_group
-
     match type(model):
         case scvi.model.SCVI:
-            return model.differential_expression(
-                idx1=cell_idx1, idx2=cell_idx2, batch_correction=batch_correct, filter_outlier_cells=remove_outliers
+            deg_df = model.differential_expression(
+                groupby=groupby,
+                group1=current_group,
+                group2=reference_group,
+                batch_correction=batch_correct,
+                filter_outlier_cells=remove_outliers,
             )
         case scvi.model.TOTALVI:
             # need to change this when/if we add support for mudata objects.  It appears that TotalVI expects that
             # adata.X has gene expression while adata.obsm[protein_key] has the protein expression
             # if we do it correctly instead of how we're doing it here, we can get protein/gene correlations
-            deg_df = model.differential_expression(idx1=cell_idx1, idx2=cell_idx2, batch_correction=batch_correct)
-            deg_df = deg_df[deg_df.index.str.endswith("x")]
-            deg_df.index = deg_df.index.str.replace("_x", "")
-            return deg_df
+            deg_df = model.differential_expression(
+                groupby=groupby, group1=current_group, group2=reference_group, batch_correction=batch_correct
+            )
         case _:
             msg = "Model is of an unrecognized type.  Currently only scVI and TotalVI models work."
             raise ValueError(msg)
+
+    if ("group1" in deg_df.columns) and (deg_df["group1"].nunique() > 1):
+        return {i: deg_df.loc[deg_df["group1"] == i, :] for i in deg_df["group1"].unique()}
+    else:
+        return deg_df
+
+
+def is_integer_array(arr: npt.ArrayLike | csr_matrix) -> bool:
+    if isinstance(arr, csr_matrix):
+        return not ((arr.toarray() % 1) != 0).any()
+    else:
+        return not ((arr % 1) != 0).any()